[#3134] comparison of integers of different signs

+++ server/api/src/rsReadCollection.cpp
@@ -18,7 +18,7 @@ rsReadCollection( rsComm_t*, int *handleInxInp,
 
     int handleInx = *handleInxInp;
 
-    if ( handleInx < 0 || handleInx >= CollHandle.size() ||
+    if ( handleInx < 0 || static_cast<std::size_t>(handleInx) >= CollHandle.size() ||
             CollHandle[handleInx].inuseFlag != FD_INUSE ) {
         rodsLog( LOG_NOTICE,
                  "rsReadCollection: handleInx %d out of range",
@@ -38,4 +38,3 @@ rsReadCollection( rsComm_t*, int *handleInxInp,
 
     return status;
 }
-
